CardのidをCardEnumに変更

+++ include/tsge/game_state/cards.hpp
@@ -5,8 +5,7 @@
 class DuckAndCover final : public Card {
  public:
   DuckAndCover()
-      : Card(static_cast<int>(CardEnum::DuckAndCover), "DuckAndCover", 3,
-             Side::USA, false) {}
+      : Card(CardEnum::DuckAndCover, "DuckAndCover", 3, Side::USA, false) {}
   [[nodiscard]]
   std::vector<CommandPtr> event(Side side) const override;
   [[nodiscard]]
@@ -22,8 +21,7 @@ class DuckAndCover final : public Card {
 
 class Fidel final : public Card {
  public:
-  Fidel()
-      : Card(static_cast<int>(CardEnum::Fidel), "Fidel", 3, Side::USSR, true) {}
+  Fidel() : Card(CardEnum::Fidel, "Fidel", 3, Side::USSR, true) {}
   [[nodiscard]]
   std::vector<CommandPtr> event(Side side) const override;
   [[nodiscard]]
@@ -33,8 +31,8 @@ class Fidel final : public Card {
 class NuclearTestBan final : public Card {
  public:
   NuclearTestBan()
-      : Card(static_cast<int>(CardEnum::NuclearTestBan), "NuclearTestBan", 4,
-             Side::NEUTRAL, false) {}
+      : Card(CardEnum::NuclearTestBan, "NuclearTestBan", 4, Side::NEUTRAL,
+             false) {}
   [[nodiscard]]
   std::vector<CommandPtr> event(Side side) const override;
   [[nodiscard]]
